requirements.txt
python-telegram-bot==20.6
httpx==0.25.2
APScheduler==3.10.4
requests==2.31.0
certifi==2023.11.17
